+++ test/czech_post_b2b_client_test.rb
@@ -10,7 +10,7 @@ class CzechPostB2bClientTest < Minitest::Test
   def test_knows_available_templates
     all_templates = CzechPostB2bClient::PrintingTemplates.all_classes.to_a
 
-    assert_equal 38, all_templates.size
+    assert_equal 34, all_templates.size
 
     template_61 = all_templates.detect { |t| t.id == 61 }
     assert_equal 'CP72 - cenný balík do zahraničí : 2x', template_61.description